+++ demo/index.d.ts
@@ -1,14 +1,12 @@
-import * as asepriteAtlas from 'aseprite-atlas'
+import * as AsepriteAtlas from 'aseprite-atlas'
 
-declare global {
-  // https://github.com/Microsoft/TypeScript/issues/4336
-  export {Animator} from asepriteAtlas
-  export {Parser} from asepriteAtlas
-  export {Aseprite} from asepriteAtlas
-  export {Atlas} from asepriteAtlas
-  export {Integer} from asepriteAtlas
-  export {Milliseconds} from asepriteAtlas
-  export {Rect} from asepriteAtlas
-  export {WH} from asepriteAtlas
-  export {XY} from asepriteAtlas
-}
+// This takes the place of re-exporting each export:
+//
+//   export {Animator} from 'aseprite-atlas'
+//   export {Parser} from 'aseprite-atlas'
+//   ...
+export = AsepriteAtlas
+
+// Everything in this module is under a variable that matches the
+// `webpack.Configuration.output.library` name.
+export as namespace AsepriteAtlas