+++ demo/index.js
@@ -7,31 +7,36 @@ if (!context) throw new Error('Missing context')
 // Use nearest neighbor scaling.
 context.imageSmoothingEnabled = false
 
-/** @type {Animator} */
+/** @type {AsepriteAtlas.Animator} */
 let animator = {period: 0, exposure: 0}
 
-/** @arg {Aseprite.File} asepriteJSON
-    @arg {HTMLImageElement} atlasImage */
-function onLoaded(asepriteJSON, atlasImage) {
-  const atlas = Parser.parse(asepriteJSON)
+/**
+ * @arg {AsepriteAtlas.Aseprite.File} asepriteJSON
+ * @arg {HTMLImageElement} atlasImage
+ */
+const onLoaded = (asepriteJSON, atlasImage) => {
+  const atlas = AsepriteAtlas.Parser.parse(asepriteJSON)
   window.requestAnimationFrame(now => loop(atlas, atlasImage, now, now))
 }
 
-/** @arg {Atlas} atlas
-    @arg {HTMLImageElement} atlasImage
-    @arg {number} then Fractional milliseconds.
-    @arg {number} now Fractional milliseconds. */
-function loop(atlas, atlasImage, then, now) {
+/**
+ * @arg {AsepriteAtlas.Atlas} atlas
+ * @arg {HTMLImageElement} atlasImage
+ * @arg {number} then Fractional milliseconds.
+ * @arg {number} now Fractional milliseconds.
+ */
+const loop = (atlas, atlasImage, then, now) => {
   const milliseconds = now - then
 
-  const animation = atlas.animations['backpacker-walkRight']
-  animator = Animator.animate(
+  const animation = /** @type {AsepriteAtlas.Atlas.Animation} */ (atlas
+    .animations['backpacker-walkRight'])
+  animator = AsepriteAtlas.Animator.animate(
     animator.period,
     animator.exposure + milliseconds,
     animation
   )
-  const index = Animator.index(animator.period, animation.cels)
-  const cel = animation.cels[index]
+  const index = AsepriteAtlas.Animator.index(animator.period, animation.cels)
+  const cel = /** @type {AsepriteAtlas.Atlas.Cel} */ (animation.cels[index])
   const scaledSize = {w: animation.size.w * 10, h: animation.size.h * 10}
 
   context.clearRect(0, 0, canvas.width, canvas.height)
@@ -50,9 +55,11 @@ function loop(atlas, atlasImage, then, now) {
   window.requestAnimationFrame(then => loop(atlas, atlasImage, now, then))
 }
 
-/** @arg {string} uri
-    @return {Promise<HTMLImageElement>} */
-function loadImage(uri) {
+/**
+ * @arg {string} uri
+ * @return {Promise<HTMLImageElement>}
+ */
+const loadImage = uri => {
   return new Promise((resolve, reject) => {
     const image = new Image()
     image.onload = () => resolve(image)
@@ -61,11 +68,11 @@ function loadImage(uri) {
   })
 }
 
-/** @arg {string} uri
-    @return {Promise<JSON>} */
-function loadJSON(uri) {
-  return fetch(uri).then(rsp => rsp.json())
-}
+/**
+ * @arg {string} uri
+ * @return {Promise<AsepriteAtlas.Aseprite.File>}
+ */
+const loadJSON = uri => fetch(uri).then(response => response.json())
 
 Promise.all([
   loadJSON('/demo/atlas.json'),