+++ webpack.config.js
@@ -1,4 +1,5 @@
 const {CleanWebpackPlugin} = require('clean-webpack-plugin')
+const path = require('path')
 const webpack = require('webpack')
 
 /** @return {webpack.Configuration} */
@@ -7,10 +8,14 @@ module.exports = () => {
     stats: 'errors-warnings',
     resolve: {extensions: ['.js', '.ts']},
     output: {
+      library: 'AsepriteAtlas',
       libraryTarget: 'umd',
       // https://github.com/webpack/webpack/issues/6525
-      globalObject: 'this'
+      globalObject: 'this',
+      path: path.resolve(__dirname, 'dist'),
+      filename: 'index.js'
     },
+    devtool: 'source-map',
     module: {rules: [{test: /\.ts$/, use: 'ts-loader'}]},
     plugins: [new CleanWebpackPlugin()]
   }